Refactor the AsyncContext
