+++ arwen/contexts/async.go
@@ -2,7 +2,6 @@ package contexts
 
 import (
 	"encoding/json"
-	"math"
 	"math/big"
 
 	"github.com/ElrondNetwork/arwen-wasm-vm/arwen"
@@ -30,6 +29,40 @@ func NewAsyncContext(host arwen.VMHost) *asyncContext {
 		ReturnData:      nil,
 		AsyncCallGroups: make([]*arwen.AsyncCallGroup, 0),
 	}
+}
+
+func (context *asyncContext) InitState() {
+	context.CallerAddr = make([]byte, 0)
+	context.ReturnData = make([]byte, 0)
+	context.AsyncCallGroups = make([]*arwen.AsyncCallGroup, 0)
+}
+
+func (context *asyncContext) SetCaller(caller []byte) {
+	context.CallerAddr = caller
+}
+
+func (context *asyncContext) PushState() {
+	// TODO the call groups must be cloned, not just referenced
+	newState := &asyncContext{
+		CallerAddr:      context.CallerAddr,
+		ReturnData:      context.ReturnData,
+		AsyncCallGroups: context.AsyncCallGroups,
+	}
+
+	context.stateStack = append(context.stateStack, newState)
+}
+
+func (context *asyncContext) PopDiscard() {
+}
+
+func (context *asyncContext) PopSetActiveState() {
+}
+
+func (context *asyncContext) PopMergeActiveState() {
+}
+
+func (context *asyncContext) ClearStateStack() {
+	context.stateStack = make([]*asyncContext, 0)
 }
 
 func (context *asyncContext) GetCallerAddress() []byte {
@@ -97,6 +130,12 @@ func (context *asyncContext) AddCall(groupID string, call *arwen.AsyncCall) erro
 		context.AddCallGroup(group)
 	}
 
+	execMode, err := context.determineExecutionMode(call.Destination, call.Data)
+	if err != nil {
+		return err
+	}
+
+	call.ExecutionMode = execMode
 	group.AddAsyncCall(call)
 
 	return nil
@@ -138,62 +177,41 @@ func (context *asyncContext) UpdateCurrentCallStatus() (*arwen.AsyncCall, error)
 	return call, nil
 }
 
-// GetPendingOnly creates a new asyncContext containing only
-// the pending AsyncCallGroups, without deleting anything from the initial asyncContext
-func (context *asyncContext) GetPendingOnly() []*arwen.AsyncCallGroup {
-	pendingGroups := make([]*arwen.AsyncCallGroup, 0)
-	var pendingGroup *arwen.AsyncCallGroup
-	for _, group := range context.AsyncCallGroups {
-		pendingGroup = nil
-		for _, asyncCall := range group.AsyncCalls {
-			if asyncCall.Status != arwen.AsyncCallPending {
-				continue
-			}
+// PrepareLegacyAsyncCall builds an AsyncCall struct from its arguments, sets it as
+// the default async call and informs Wasmer to stop contract execution with BreakpointAsyncCall
+func (context *asyncContext) PrepareLegacyAsyncCall(address []byte, data []byte, value []byte) error {
+	legacyGroupID := arwen.LegacyAsyncCallGroupID
 
-			if pendingGroup == nil {
-				pendingGroup = arwen.NewAsyncCallGroup(group.Identifier)
-				pendingGroups = append(pendingGroups, pendingGroup)
-			}
-
-			pendingGroup.AsyncCalls = append(pendingGroup.AsyncCalls, asyncCall)
-		}
+	_, exists := context.GetCallGroup(legacyGroupID)
+	if exists {
+		return arwen.ErrOnlyOneLegacyAsyncCallAllowed
 	}
 
-	return pendingGroups
-}
-
-func (context *asyncContext) InitState() {
-	context.CallerAddr = make([]byte, 0)
-	context.ReturnData = make([]byte, 0)
-	context.AsyncCallGroups = make([]*arwen.AsyncCallGroup, 0)
-}
-
-func (context *asyncContext) SetCaller(caller []byte) {
-	context.CallerAddr = caller
-}
-
-func (context *asyncContext) PushState() {
-	// TODO the call groups must be cloned, not just referenced
-	newState := &asyncContext{
-		CallerAddr:      context.CallerAddr,
-		ReturnData:      context.ReturnData,
-		AsyncCallGroups: context.AsyncCallGroups,
+	gasToLock, err := context.prepareGasForAsyncCall()
+	if err != nil {
+		return err
 	}
 
-	context.stateStack = append(context.stateStack, newState)
-}
+	metering := context.host.Metering()
+	gas := metering.GasLeft()
 
-func (context *asyncContext) PopDiscard() {
-}
+	err = context.AddCall(legacyGroupID, &arwen.AsyncCall{
+		Status:          arwen.AsyncCallPending,
+		Destination:     address,
+		Data:            data,
+		ValueBytes:      value,
+		SuccessCallback: arwen.CallbackFunctionName,
+		ErrorCallback:   arwen.CallbackFunctionName,
+		ProvidedGas:     gas,
+		GasLocked:       gasToLock,
+	})
+	if err != nil {
+		return err
+	}
 
-func (context *asyncContext) PopSetActiveState() {
-}
+	context.host.Runtime().SetRuntimeBreakpointValue(arwen.BreakpointAsyncCall)
 
-func (context *asyncContext) PopMergeActiveState() {
-}
-
-func (context *asyncContext) ClearStateStack() {
-	context.stateStack = make([]*asyncContext, 0)
+	return nil
 }
 
 // Execute is the entry-point of the async calling mechanism; it is called by
@@ -210,6 +228,10 @@ func (context *asyncContext) ClearStateStack() {
 // because synchronous AsyncCalls are executed with
 // host.ExecuteOnDestContext(), which, in turn, calls asyncContext.Execute() to
 // resolve AsyncCalls generated by the AsyncCalls, and so on.
+
+// Moreover, host.ExecuteOnDestContext() will push the state stack of the
+// AsyncContext and work with a clean state before calling Execute(), making
+// Execute() and host.ExecuteOnDestContext() mutually reentrant.
 func (context *asyncContext) Execute() error {
 	if context.IsComplete() {
 		return nil
@@ -222,16 +244,9 @@ func (context *asyncContext) Execute() error {
 		return err
 	}
 
-	for groupIndex, group := range context.AsyncCallGroups {
-		// Execute the call group strictly synchronously (no asynchronous calls allowed)
-		err := context.executeCallGroup(group, true)
-		if err != nil {
-			return err
-		}
-
-		if group.IsCompleted() {
-			context.DeleteCallGroup(groupIndex)
-		}
+	err = context.executeSynchronousCalls()
+	if err != nil {
+		return err
 	}
 
 	// Step 2: redistribute unspent gas; then, in one combined step, do the
@@ -246,10 +261,11 @@ func (context *asyncContext) Execute() error {
 	}
 
 	for _, group := range context.AsyncCallGroups {
-		// Execute the call group allowing asynchronous (cross-shard) calls as well
-		err = context.executeCallGroup(group, false)
-		if err != nil {
-			return err
+		for _, call := range group.AsyncCalls {
+			err = context.executeAsyncCall(call)
+			if err != nil {
+				return err
+			}
 		}
 	}
 
@@ -263,66 +279,49 @@ func (context *asyncContext) Execute() error {
 	return nil
 }
 
-// PrepareLegacyAsyncCall builds an AsyncCall struct from its arguments, sets it as
-// the default async call and informs Wasmer to stop contract execution with BreakpointAsyncCall
-func (context *asyncContext) PrepareLegacyAsyncCall(address []byte, data []byte, value []byte) error {
-	legacyGroupID := arwen.LegacyAsyncCallGroupID
-	legacyCallback := []byte(arwen.CallbackFunctionName)
-
-	_, exists := context.GetCallGroup(legacyGroupID)
-	if exists {
-		return arwen.ErrOnlyOneLegacyAsyncCallAllowed
+// executeAsyncCall executes asynchronous cross-shard calls to built-in
+// functions and contracts
+func (context *asyncContext) executeAsyncCall(asyncCall *arwen.AsyncCall) error {
+	if asyncCall.ExecutionMode == arwen.AsyncUnknown {
+		return context.sendAsyncCallCrossShard(asyncCall)
 	}
 
-	err := context.CreateAndAddCall(legacyGroupID,
-		address,
-		data,
-		value,
-		legacyCallback,
-		legacyCallback,
-		math.MaxUint64,
-	)
-	if err != nil {
-		return err
-	}
+	if asyncCall.ExecutionMode == arwen.AsyncBuiltinFunc {
+		// Built-in functions will handle cross-shard calls themselves, by
+		// generating entries in vmOutput.OutputAccounts, but they need to be
+		// executed synchronously to do that. It is not necessary to call
+		// sendAsyncCallCrossShard(). The vmOutput produced by the built-in
+		// function, containing the cross-shard call, has ALREADY been merged into
+		// the main output by the inner call to host.ExecuteOnDestContext().  The
+		// status of the AsyncCall is not updated here - it will be updated by
+		// PostprocessCrossShardCallback(), when the cross-shard call returns.
+		destinationCallInput, err := context.createSyncCallInput(asyncCall)
+		if err != nil {
+			return err
+		}
 
-	context.host.Runtime().SetRuntimeBreakpointValue(arwen.BreakpointAsyncCall)
+		vmOutput, err := context.host.ExecuteOnDestContext(destinationCallInput)
+		if err != nil {
+			return err
+		}
+
+		// If the synchronous half of the built-in function call has failed, go no
+		// further and execute the error callback of this AsyncCall.
+		if vmOutput.ReturnCode != vmcommon.Ok {
+			asyncCall.UpdateStatus(vmOutput.ReturnCode)
+			callbackVMOutput, callbackErr := context.executeSyncCallback(asyncCall, vmOutput, err)
+			context.finishSyncExecution(callbackVMOutput, callbackErr)
+		}
+
+		return nil
+	}
 
 	return nil
 }
 
-// CreateAndAddAsyncCall creates a new AsyncCall from its arguments and adds it
-// to the specified group
-func (context *asyncContext) CreateAndAddCall(
-	groupID string,
-	address []byte,
-	data []byte,
-	value []byte,
-	successCallback []byte,
-	errorCallback []byte,
-	gas uint64,
-) error {
-
-	gasToLock, err := context.prepareGasForAsyncCall()
-	if err != nil {
-		return err
-	}
-
-	if gas == math.MaxUint64 {
-		metering := context.host.Metering()
-		gas = metering.GasLeft()
-	}
-
-	return context.AddCall(groupID, &arwen.AsyncCall{
-		Status:          arwen.AsyncCallPending,
-		Destination:     address,
-		Data:            data,
-		ValueBytes:      value,
-		SuccessCallback: string(successCallback),
-		ErrorCallback:   string(errorCallback),
-		ProvidedGas:     gas,
-		GasLocked:       gasToLock,
-	})
+func (context *asyncContext) executeAsyncCallGroupCallback(group *arwen.AsyncCallGroup) error {
+	// TODO implement this
+	return nil
 }
 
 func (context *asyncContext) prepareGasForAsyncCall() (uint64, error) {
@@ -391,341 +390,27 @@ func (context *asyncContext) PostprocessCrossShardCallback() error {
 		return nil
 	}
 
+	// The current group expects no more callbacks, so its own callback can be
+	// executed now.
+	err = context.executeAsyncCallGroupCallback(currentCallGroup)
+	if err != nil {
+		return err
+	}
+
 	context.DeleteCallGroupByID(currentGroupID)
 	// Are we still waiting for callbacks to return?
 	if context.HasPendingCallGroups() {
 		return nil
 	}
 
+	// There are no more callbacks to return from other shards. The context can
+	// be deleted from storage.
 	err = context.Delete()
 	if err != nil {
 		return err
 	}
 
 	return context.executeAsyncContextCallback()
-}
-
-// executeAsyncContextCallback will either execute a sync call (in-shard) to
-// the original caller by invoking its callback directly, or will dispatch a
-// cross-shard callback to it.
-func (context *asyncContext) executeAsyncContextCallback() error {
-	execMode, err := context.DetermineExecutionMode(context.CallerAddr, context.ReturnData)
-	if err != nil {
-		return err
-	}
-
-	if execMode != arwen.SyncExecution {
-		return context.sendContextCallbackToOriginalCaller()
-	}
-
-	// The caller is in the same shard, execute its callback
-	callbackCallInput := context.createSyncContextCallbackInput()
-
-	callbackVMOutput, callBackErr := context.host.ExecuteOnDestContext(callbackCallInput)
-	context.finishSyncExecution(callbackVMOutput, callBackErr)
-
-	return nil
-}
-
-func (context *asyncContext) sendContextCallbackToOriginalCaller() error {
-	host := context.host
-	runtime := host.Runtime()
-	output := host.Output()
-	metering := host.Metering()
-	currentCall := runtime.GetVMInput()
-
-	err := output.Transfer(
-		context.CallerAddr,
-		runtime.GetSCAddress(),
-		metering.GasLeft(),
-		0,
-		currentCall.CallValue,
-		context.ReturnData,
-		vmcommon.AsynchronousCallBack,
-	)
-	if err != nil {
-		metering.UseGas(metering.GasLeft())
-		runtime.FailExecution(err)
-		return err
-	}
-
-	return nil
-}
-
-func (context *asyncContext) createSyncContextCallbackInput() *vmcommon.ContractCallInput {
-	host := context.host
-	runtime := host.Runtime()
-	metering := host.Metering()
-
-	_, arguments, err := host.CallArgsParser().ParseData(string(context.ReturnData))
-	if err != nil {
-		arguments = [][]byte{context.ReturnData}
-	}
-
-	// TODO ensure a new value for VMInput.CurrentTxHash
-	input := &vmcommon.ContractCallInput{
-		VMInput: vmcommon.VMInput{
-			CallerAddr:     runtime.GetSCAddress(),
-			Arguments:      arguments,
-			CallValue:      runtime.GetVMInput().CallValue,
-			CallType:       vmcommon.AsynchronousCallBack,
-			GasPrice:       runtime.GetVMInput().GasPrice,
-			GasProvided:    metering.GasLeft(),
-			CurrentTxHash:  runtime.GetCurrentTxHash(),
-			OriginalTxHash: runtime.GetOriginalTxHash(),
-			PrevTxHash:     runtime.GetPrevTxHash(),
-		},
-		RecipientAddr: context.CallerAddr,
-		Function:      arwen.CallbackFunctionName, // TODO currently default; will customize in AsynContext
-	}
-	return input
-}
-
-func (context *asyncContext) executeCallGroup(
-	group *arwen.AsyncCallGroup,
-	syncExecutionOnly bool,
-) error {
-	for _, asyncCall := range group.AsyncCalls {
-		err := context.executeCall(asyncCall, syncExecutionOnly)
-		if err != nil {
-			return err
-		}
-	}
-
-	group.DeleteCompletedAsyncCalls()
-
-	// If ALL the AsyncCalls in the AsyncCallGroup were executed synchronously,
-	// then the AsyncCallGroup can have its callback executed.
-	if group.IsCompleted() {
-		// TODO reenable this, after allowing a gas limit for it and deciding what
-		// arguments it receives (this method is currently a NOP and returns nil)
-		return context.executeAsyncCallGroupCallback(group)
-	}
-
-	return nil
-}
-
-// TODO split into two different functions, for sync execution and async
-// execution, and remove parameter syncExecutionOnly
-func (context *asyncContext) executeCall(
-	asyncCall *arwen.AsyncCall,
-	syncExecutionOnly bool,
-) error {
-	execMode, err := context.DetermineExecutionMode(asyncCall.Destination, asyncCall.Data)
-	if err != nil {
-		return err
-	}
-
-	if execMode == arwen.SyncExecution {
-		vmOutput, err := context.executeSyncCall(asyncCall)
-
-		// The vmOutput instance returned by host.executeSyncCall() is never nil,
-		// by design. Using it without checking for err is safe here.
-		asyncCall.UpdateStatus(vmOutput.ReturnCode)
-
-		// TODO host.executeSyncCallback() returns a vmOutput produced by executing
-		// the callback. Information from this vmOutput should be preserved in the
-		// pending AsyncCallGroup, and made available to the callback of the
-		// AsyncCallGroup (currently not implemented).
-		callbackVMOutput, callbackErr := context.executeSyncCallback(asyncCall, vmOutput, err)
-		context.finishSyncExecution(callbackVMOutput, callbackErr)
-		return nil
-	}
-
-	if syncExecutionOnly {
-		return nil
-	}
-
-	if execMode == arwen.AsyncBuiltinFunc {
-		// Built-in functions will handle cross-shard calls themselves, by
-		// generating entries in vmOutput.OutputAccounts, but they need to be
-		// executed synchronously to do that. It is not necessary to call
-		// sendAsyncCallCrossShard(). The vmOutput produced by the built-in
-		// function, containing the cross-shard call, has ALREADY been merged into
-		// the main output by the inner call to host.ExecuteOnDestContext().  The
-		// status of the AsyncCall is not updated here - it will be updated by
-		// postprocessCrossShardCallback(), when the cross-shard call returns.
-		vmOutput, err := context.executeSyncCall(asyncCall)
-		if err != nil {
-			return err
-		}
-
-		if vmOutput.ReturnCode != vmcommon.Ok {
-			asyncCall.UpdateStatus(vmOutput.ReturnCode)
-			callbackVMOutput, callbackErr := context.executeSyncCallback(asyncCall, vmOutput, err)
-			context.finishSyncExecution(callbackVMOutput, callbackErr)
-		}
-
-		return nil
-	}
-
-	if execMode == arwen.AsyncUnknown {
-		return context.sendAsyncCallCrossShard(asyncCall)
-	}
-
-	return nil
-}
-
-func (context *asyncContext) executeSyncCall(asyncCall *arwen.AsyncCall) (*vmcommon.VMOutput, error) {
-	destinationCallInput, err := context.createSyncCallInput(asyncCall)
-	if err != nil {
-		return nil, err
-	}
-
-	return context.host.ExecuteOnDestContext(destinationCallInput)
-}
-
-func (context *asyncContext) executeSyncCallback(
-	asyncCall *arwen.AsyncCall,
-	vmOutput *vmcommon.VMOutput,
-	err error,
-) (*vmcommon.VMOutput, error) {
-
-	callbackInput, err := context.createSyncCallbackInput(asyncCall, vmOutput, err)
-	if err != nil {
-		return nil, err
-	}
-
-	return context.host.ExecuteOnDestContext(callbackInput)
-}
-
-func (context *asyncContext) executeAsyncCallGroupCallback(group *arwen.AsyncCallGroup) error {
-	// TODO implement this
-	return nil
-}
-
-func (context *asyncContext) sendAsyncCallCrossShard(asyncCall arwen.AsyncCallHandler) error {
-	host := context.host
-	runtime := host.Runtime()
-	output := host.Output()
-
-	err := output.Transfer(
-		asyncCall.GetDestination(),
-		runtime.GetSCAddress(),
-		asyncCall.GetGasLimit(),
-		asyncCall.GetGasLocked(),
-		big.NewInt(0).SetBytes(asyncCall.GetValueBytes()),
-		asyncCall.GetData(),
-		vmcommon.AsynchronousCall,
-	)
-	if err != nil {
-		metering := host.Metering()
-		metering.UseGas(metering.GasLeft())
-		runtime.FailExecution(err)
-		return err
-	}
-
-	return nil
-}
-
-func (context *asyncContext) createSyncCallInput(asyncCall arwen.AsyncCallHandler) (*vmcommon.ContractCallInput, error) {
-	host := context.host
-	runtime := host.Runtime()
-	sender := runtime.GetSCAddress()
-
-	function, arguments, err := host.CallArgsParser().ParseData(string(asyncCall.GetData()))
-	if err != nil {
-		return nil, err
-	}
-
-	gasLimit := asyncCall.GetGasLimit()
-	gasToUse := host.Metering().GasSchedule().ElrondAPICost.AsyncCallStep
-	if gasLimit <= gasToUse {
-		return nil, arwen.ErrNotEnoughGas
-	}
-	gasLimit -= gasToUse
-
-	contractCallInput := &vmcommon.ContractCallInput{
-		VMInput: vmcommon.VMInput{
-			CallerAddr:     sender,
-			Arguments:      arguments,
-			CallValue:      big.NewInt(0).SetBytes(asyncCall.GetValueBytes()),
-			CallType:       vmcommon.AsynchronousCall,
-			GasPrice:       runtime.GetVMInput().GasPrice,
-			GasProvided:    gasLimit,
-			CurrentTxHash:  runtime.GetCurrentTxHash(),
-			OriginalTxHash: runtime.GetOriginalTxHash(),
-			PrevTxHash:     runtime.GetPrevTxHash(),
-		},
-		RecipientAddr: asyncCall.GetDestination(),
-		Function:      function,
-	}
-
-	return contractCallInput, nil
-}
-
-func (context *asyncContext) createSyncCallbackInput(
-	asyncCall *arwen.AsyncCall,
-	vmOutput *vmcommon.VMOutput,
-	destinationErr error,
-) (*vmcommon.ContractCallInput, error) {
-	metering := context.host.Metering()
-	runtime := context.host.Runtime()
-
-	// always provide return code as the first argument to callback function
-	arguments := [][]byte{
-		big.NewInt(int64(vmOutput.ReturnCode)).Bytes(),
-	}
-	if destinationErr == nil {
-		// when execution went Ok, callBack arguments are:
-		// [0, result1, result2, ....]
-		arguments = append(arguments, vmOutput.ReturnData...)
-	} else {
-		// when execution returned error, callBack arguments are:
-		// [error code, error message]
-		arguments = append(arguments, []byte(vmOutput.ReturnMessage))
-	}
-
-	callbackFunction := asyncCall.GetCallbackName()
-
-	gasLimit := vmOutput.GasRemaining + asyncCall.GetGasLocked()
-	dataLength := computeDataLengthFromArguments(callbackFunction, arguments)
-
-	gasToUse := metering.GasSchedule().ElrondAPICost.AsyncCallStep
-	gasToUse += metering.GasSchedule().BaseOperationCost.DataCopyPerByte * uint64(dataLength)
-	if gasLimit <= gasToUse {
-		return nil, arwen.ErrNotEnoughGas
-	}
-	gasLimit -= gasToUse
-
-	// Return to the sender SC, calling its specified callback method.
-	contractCallInput := &vmcommon.ContractCallInput{
-		VMInput: vmcommon.VMInput{
-			CallerAddr:     asyncCall.Destination,
-			Arguments:      arguments,
-			CallValue:      big.NewInt(0),
-			CallType:       vmcommon.AsynchronousCallBack,
-			GasPrice:       runtime.GetVMInput().GasPrice,
-			GasProvided:    gasLimit,
-			CurrentTxHash:  runtime.GetCurrentTxHash(),
-			OriginalTxHash: runtime.GetOriginalTxHash(),
-			PrevTxHash:     runtime.GetPrevTxHash(),
-		},
-		RecipientAddr: runtime.GetSCAddress(),
-		Function:      callbackFunction,
-	}
-
-	return contractCallInput, nil
-}
-
-func (context *asyncContext) finishSyncExecution(vmOutput *vmcommon.VMOutput, err error) {
-	if err == nil {
-		return
-	}
-
-	runtime := context.host.Runtime()
-	output := context.host.Output()
-
-	runtime.GetVMInput().GasProvided = 0
-
-	if vmOutput == nil {
-		vmOutput = output.CreateVMOutputInCaseOfError(err)
-	}
-
-	output.SetReturnMessage(vmOutput.ReturnMessage)
-
-	output.Finish([]byte(vmOutput.ReturnCode.String()))
-	output.Finish(runtime.GetCurrentTxHash())
 }
 
 func (context *asyncContext) HasPendingCallGroups() bool {
@@ -789,7 +474,7 @@ func (context *asyncContext) Delete() error {
 	return err
 }
 
-func (context *asyncContext) DetermineExecutionMode(destination []byte, data []byte) (arwen.AsyncCallExecutionMode, error) {
+func (context *asyncContext) determineExecutionMode(destination []byte, data []byte) (arwen.AsyncCallExecutionMode, error) {
 	runtime := context.host.Runtime()
 	blockchain := context.host.Blockchain()
 
@@ -856,6 +541,77 @@ func (context *asyncContext) setupAsyncCallsGas() error {
 				asyncCall.GasLimit = gasShare
 			}
 		}
+	}
+
+	return nil
+}
+
+func (context *asyncContext) sendAsyncCallCrossShard(asyncCall arwen.AsyncCallHandler) error {
+	host := context.host
+	runtime := host.Runtime()
+	output := host.Output()
+
+	err := output.Transfer(
+		asyncCall.GetDestination(),
+		runtime.GetSCAddress(),
+		asyncCall.GetGasLimit(),
+		asyncCall.GetGasLocked(),
+		big.NewInt(0).SetBytes(asyncCall.GetValueBytes()),
+		asyncCall.GetData(),
+		vmcommon.AsynchronousCall,
+	)
+	if err != nil {
+		metering := host.Metering()
+		metering.UseGas(metering.GasLeft())
+		runtime.FailExecution(err)
+		return err
+	}
+
+	return nil
+}
+
+// executeAsyncContextCallback will either execute a sync call (in-shard) to
+// the original caller by invoking its callback directly, or will dispatch a
+// cross-shard callback to it.
+func (context *asyncContext) executeAsyncContextCallback() error {
+	execMode, err := context.determineExecutionMode(context.CallerAddr, context.ReturnData)
+	if err != nil {
+		return err
+	}
+
+	if execMode != arwen.SyncExecution {
+		return context.sendContextCallbackToOriginalCaller()
+	}
+
+	// The caller is in the same shard, execute its callback
+	callbackCallInput := context.createSyncContextCallbackInput()
+
+	callbackVMOutput, callBackErr := context.host.ExecuteOnDestContext(callbackCallInput)
+	context.finishSyncExecution(callbackVMOutput, callBackErr)
+
+	return nil
+}
+
+func (context *asyncContext) sendContextCallbackToOriginalCaller() error {
+	host := context.host
+	runtime := host.Runtime()
+	output := host.Output()
+	metering := host.Metering()
+	currentCall := runtime.GetVMInput()
+
+	err := output.Transfer(
+		context.CallerAddr,
+		runtime.GetSCAddress(),
+		metering.GasLeft(),
+		0,
+		currentCall.CallValue,
+		context.ReturnData,
+		vmcommon.AsynchronousCallBack,
+	)
+	if err != nil {
+		metering.UseGas(metering.GasLeft())
+		runtime.FailExecution(err)
+		return err
 	}
 
 	return nil
